Fix errors & update integration tests
