+++ components/camel-hazelcast/src/main/java/org/apache/camel/integration/hazelcast/HazelcastIdempotentRepository.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.processor.idempotent.hazelcast;
+package org.apache.camel.integration.hazelcast;
 
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.map.IMap;
