+++ components/camel-hazelcast/src/test/java/org/apache/camel/integration/HazelcastSedaInOutIT.java
@@ -14,42 +14,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.hazelcast;
+package org.apache.camel.integration;
 
-import com.hazelcast.core.Hazelcast;
-import com.hazelcast.core.HazelcastInstance;
-import org.apache.camel.CamelContext;
 import org.apache.camel.EndpointInject;
 import org.apache.camel.Exchange;
 import org.apache.camel.ExchangePattern;
 import org.apache.camel.Processor;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
-import org.apache.camel.test.junit5.CamelTestSupport;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-public class HazelcastSedaInOutTest extends CamelTestSupport {
+public class HazelcastSedaInOutIT extends BaseHazelcast {
 
     @EndpointInject("mock:result")
     private MockEndpoint mock;
-
-    private HazelcastInstance hazelcastInstance;
-
-    @BeforeAll
-    public void beforeEach() {
-        hazelcastInstance = Hazelcast.newHazelcastInstance();
-    }
-
-    @AfterAll
-    public void afterEach() {
-        if (hazelcastInstance != null) {
-            hazelcastInstance.shutdown();
-        }
-    }
 
     @Test
     public void sendInOut() throws Exception {
@@ -63,13 +43,6 @@ public class HazelcastSedaInOutTest extends CamelTestSupport {
         });
         assertMockEndpointsSatisfied();
         mock.reset();
-    }
-
-    @Override
-    protected CamelContext createCamelContext() throws Exception {
-        CamelContext context = super.createCamelContext();
-        HazelcastCamelTestHelper.registerHazelcastComponents(context, hazelcastInstance);
-        return context;
     }
 
     @Override
