+++ components/camel-hazelcast/src/test/java/org/apache/camel/processor/aggregate/hazelcast/HazelcastAggregationRepositoryOperationsIT.java
@@ -26,7 +26,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
-public class HazelcastAggregationRepositoryOperationsTest extends HazelcastAggregationRepositoryCamelTestSupport {
+public class HazelcastAggregationRepositoryOperationsIT extends HazelcastAggregationRepositoryCamelTestSupport {
 
     private static final String THREAD_SAFE_REPO = "threadSafeRepo";
     private static final String OPTIMISTIC_REPO = "optimisticRepo";
