+++ components/camel-hazelcast/src/test/java/org/apache/camel/processor/aggregate/hazelcast/HazelcastAggregationRepositoryRecoverableRoutesIT.java
@@ -23,15 +23,19 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
 
-public class HazelcastAggregationRepositoryRoutesTest extends HazelcastAggregationRepositoryCamelTestSupport {
+public class HazelcastAggregationRepositoryRecoverableRoutesIT extends HazelcastAggregationRepositoryCamelTestSupport {
 
     private static final String REPO_NAME = "routeTestRepo";
     private static final String MOCK_GOTCHA = "mock:gotcha";
+    private static final String MOCK_FAILURE = "mock:failure";
     private static final String DIRECT_ONE = "direct:one";
     private static final String DIRECT_TWO = "direct:two";
 
     @EndpointInject(MOCK_GOTCHA)
-    private MockEndpoint mock;
+    private MockEndpoint mockGotcha;
+
+    @EndpointInject(MOCK_FAILURE)
+    private MockEndpoint mockFailure;
 
     @Produce(DIRECT_ONE)
     private ProducerTemplate produceOne;
@@ -40,7 +44,7 @@ public class HazelcastAggregationRepositoryRoutesTest extends HazelcastAggregati
     private ProducerTemplate produceTwo;
 
     @Test
-    public void checkAggregationFromTwoRoutes() throws Exception {
+    public void checkAggregationFromTwoRoutesWithRecovery() throws Exception {
         final HazelcastAggregationRepository repoOne = new HazelcastAggregationRepository(REPO_NAME, false, getFirstInstance());
 
         final HazelcastAggregationRepository repoTwo
@@ -48,16 +52,28 @@ public class HazelcastAggregationRepositoryRoutesTest extends HazelcastAggregati
 
         final int completionSize = 4;
         final String correlator = "CORRELATOR";
+
         RouteBuilder rbOne = new RouteBuilder() {
             @Override
             public void configure() throws Exception {
 
-                from(DIRECT_ONE).routeId("AggregatingRouteOne")
+                onException(EverythingIsLostException.class)
+                        .handled(true)
+                        .useOriginalMessage()
+                        .to(MOCK_GOTCHA)
+                        .end();
+
+                interceptSendToEndpoint(MOCK_FAILURE)
+                        .throwException(new EverythingIsLostException("The field is lost... everything is lost"))
+                        .end();
+
+                from(DIRECT_ONE)
                         .aggregate(header(correlator))
                         .aggregationRepository(repoOne)
                         .aggregationStrategy(new SumOfIntsAggregationStrategy())
                         .completionSize(completionSize)
-                        .to(MOCK_GOTCHA);
+                        .to(MOCK_FAILURE);
+
             }
         };
 
@@ -65,12 +81,22 @@ public class HazelcastAggregationRepositoryRoutesTest extends HazelcastAggregati
             @Override
             public void configure() throws Exception {
 
-                from(DIRECT_TWO).routeId("AggregatingRouteTwo")
+                onException(EverythingIsLostException.class)
+                        .handled(true)
+                        .useOriginalMessage()
+                        .to(MOCK_GOTCHA)
+                        .end();
+
+                interceptSendToEndpoint(MOCK_FAILURE)
+                        .throwException(new EverythingIsLostException("The field is lost... everything is lost"))
+                        .end();
+
+                from(DIRECT_TWO)
                         .aggregate(header(correlator))
                         .aggregationRepository(repoTwo)
                         .aggregationStrategy(new SumOfIntsAggregationStrategy())
                         .completionSize(completionSize)
-                        .to(MOCK_GOTCHA);
+                        .to(MOCK_FAILURE);
             }
         };
 
@@ -78,68 +104,40 @@ public class HazelcastAggregationRepositoryRoutesTest extends HazelcastAggregati
         context().addRoutes(rbTwo);
         context().start();
 
-        mock.expectedMessageCount(1);
-        mock.expectedBodiesReceived(1 + 2 + 3 + 4);
+        mockFailure.expectedMessageCount(0);
+        mockGotcha.expectedMessageCount(1);
+        mockGotcha.expectedBodiesReceived(1 + 2 + 3 + 4);
 
-        produceOne.sendBodyAndHeader(1, correlator, correlator);
+        produceOne.sendBodyAndHeader(4, correlator, correlator);
         produceTwo.sendBodyAndHeader(2, correlator, correlator);
         produceOne.sendBodyAndHeader(3, correlator, correlator);
-        produceOne.sendBodyAndHeader(4, correlator, correlator);
+        produceTwo.sendBodyAndHeader(1, correlator, correlator);
 
-        mock.assertIsSatisfied();
+        mockFailure.assertIsSatisfied();
+        mockFailure.assertIsSatisfied();
     }
 
-    @Test
-    public void checkAggregationFromTwoRoutesNoRecovery() throws Exception {
-        final HazelcastAggregationRepository repoOne = new HazelcastAggregationRepository(REPO_NAME, false, getFirstInstance());
+    private static final class EverythingIsLostException extends Exception {
+        private static final long serialVersionUID = 1L;
 
-        final HazelcastAggregationRepository repoTwo
-                = new HazelcastAggregationRepository(REPO_NAME, false, getSecondInstance());
+        private EverythingIsLostException() {
+        }
 
-        repoOne.setUseRecovery(false);
-        repoTwo.setUseRecovery(false);
+        private EverythingIsLostException(String message) {
+            super(message);
+        }
 
-        final int completionSize = 4;
-        final String correlator = "CORRELATOR";
-        RouteBuilder rbOne = new RouteBuilder() {
-            @Override
-            public void configure() throws Exception {
+        private EverythingIsLostException(String message, Throwable cause) {
+            super(message, cause);
+        }
 
-                from(DIRECT_ONE).routeId("AggregatingRouteOne")
-                        .aggregate(header(correlator))
-                        .aggregationRepository(repoOne)
-                        .aggregationStrategy(new SumOfIntsAggregationStrategy())
-                        .completionSize(completionSize)
-                        .to(MOCK_GOTCHA);
-            }
-        };
+        private EverythingIsLostException(Throwable cause) {
+            super(cause);
+        }
 
-        RouteBuilder rbTwo = new RouteBuilder() {
-            @Override
-            public void configure() throws Exception {
-
-                from(DIRECT_TWO).routeId("AggregatingRouteTwo")
-                        .aggregate(header(correlator))
-                        .aggregationRepository(repoTwo)
-                        .aggregationStrategy(new SumOfIntsAggregationStrategy())
-                        .completionSize(completionSize)
-                        .to(MOCK_GOTCHA);
-            }
-        };
-
-        context().addRoutes(rbOne);
-        context().addRoutes(rbTwo);
-        context().start();
-
-        mock.expectedMessageCount(1);
-        mock.expectedBodiesReceived(1 + 2 + 3 + 4);
-
-        produceOne.sendBodyAndHeader(1, correlator, correlator);
-        produceTwo.sendBodyAndHeader(2, correlator, correlator);
-        produceOne.sendBodyAndHeader(3, correlator, correlator);
-        produceOne.sendBodyAndHeader(4, correlator, correlator);
-
-        mock.assertIsSatisfied();
+        // not in jdk6
+        //        private EverythingIsLostException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        //            super(message, cause, enableSuppression, writableStackTrace);
+        //        }
     }
-
 }
