+++ components/camel-hazelcast/src/test/java/org/apache/camel/processor/aggregate/hazelcast/HazelcastAggregationRepositoryRoutesIT.java
@@ -23,19 +23,15 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
 
-public class HazelcastAggregationRepositoryRecoverableRoutesTest extends HazelcastAggregationRepositoryCamelTestSupport {
+public class HazelcastAggregationRepositoryRoutesIT extends HazelcastAggregationRepositoryCamelTestSupport {
 
     private static final String REPO_NAME = "routeTestRepo";
     private static final String MOCK_GOTCHA = "mock:gotcha";
-    private static final String MOCK_FAILURE = "mock:failure";
     private static final String DIRECT_ONE = "direct:one";
     private static final String DIRECT_TWO = "direct:two";
 
     @EndpointInject(MOCK_GOTCHA)
-    private MockEndpoint mockGotcha;
-
-    @EndpointInject(MOCK_FAILURE)
-    private MockEndpoint mockFailure;
+    private MockEndpoint mock;
 
     @Produce(DIRECT_ONE)
     private ProducerTemplate produceOne;
@@ -44,7 +40,7 @@ public class HazelcastAggregationRepositoryRecoverableRoutesTest extends Hazelca
     private ProducerTemplate produceTwo;
 
     @Test
-    public void checkAggregationFromTwoRoutesWithRecovery() throws Exception {
+    public void checkAggregationFromTwoRoutes() throws Exception {
         final HazelcastAggregationRepository repoOne = new HazelcastAggregationRepository(REPO_NAME, false, getFirstInstance());
 
         final HazelcastAggregationRepository repoTwo
@@ -52,28 +48,16 @@ public class HazelcastAggregationRepositoryRecoverableRoutesTest extends Hazelca
 
         final int completionSize = 4;
         final String correlator = "CORRELATOR";
-
         RouteBuilder rbOne = new RouteBuilder() {
             @Override
             public void configure() throws Exception {
 
-                onException(EverythingIsLostException.class)
-                        .handled(true)
-                        .useOriginalMessage()
-                        .to(MOCK_GOTCHA)
-                        .end();
-
-                interceptSendToEndpoint(MOCK_FAILURE)
-                        .throwException(new EverythingIsLostException("The field is lost... everything is lost"))
-                        .end();
-
-                from(DIRECT_ONE)
+                from(DIRECT_ONE).routeId("AggregatingRouteOne")
                         .aggregate(header(correlator))
                         .aggregationRepository(repoOne)
                         .aggregationStrategy(new SumOfIntsAggregationStrategy())
                         .completionSize(completionSize)
-                        .to(MOCK_FAILURE);
-
+                        .to(MOCK_GOTCHA);
             }
         };
 
@@ -81,22 +65,12 @@ public class HazelcastAggregationRepositoryRecoverableRoutesTest extends Hazelca
             @Override
             public void configure() throws Exception {
 
-                onException(EverythingIsLostException.class)
-                        .handled(true)
-                        .useOriginalMessage()
-                        .to(MOCK_GOTCHA)
-                        .end();
-
-                interceptSendToEndpoint(MOCK_FAILURE)
-                        .throwException(new EverythingIsLostException("The field is lost... everything is lost"))
-                        .end();
-
-                from(DIRECT_TWO)
+                from(DIRECT_TWO).routeId("AggregatingRouteTwo")
                         .aggregate(header(correlator))
                         .aggregationRepository(repoTwo)
                         .aggregationStrategy(new SumOfIntsAggregationStrategy())
                         .completionSize(completionSize)
-                        .to(MOCK_FAILURE);
+                        .to(MOCK_GOTCHA);
             }
         };
 
@@ -104,40 +78,68 @@ public class HazelcastAggregationRepositoryRecoverableRoutesTest extends Hazelca
         context().addRoutes(rbTwo);
         context().start();
 
-        mockFailure.expectedMessageCount(0);
-        mockGotcha.expectedMessageCount(1);
-        mockGotcha.expectedBodiesReceived(1 + 2 + 3 + 4);
+        mock.expectedMessageCount(1);
+        mock.expectedBodiesReceived(1 + 2 + 3 + 4);
 
-        produceOne.sendBodyAndHeader(4, correlator, correlator);
+        produceOne.sendBodyAndHeader(1, correlator, correlator);
         produceTwo.sendBodyAndHeader(2, correlator, correlator);
         produceOne.sendBodyAndHeader(3, correlator, correlator);
-        produceTwo.sendBodyAndHeader(1, correlator, correlator);
+        produceOne.sendBodyAndHeader(4, correlator, correlator);
 
-        mockFailure.assertIsSatisfied();
-        mockFailure.assertIsSatisfied();
+        mock.assertIsSatisfied();
     }
 
-    private static final class EverythingIsLostException extends Exception {
-        private static final long serialVersionUID = 1L;
+    @Test
+    public void checkAggregationFromTwoRoutesNoRecovery() throws Exception {
+        final HazelcastAggregationRepository repoOne = new HazelcastAggregationRepository(REPO_NAME, false, getFirstInstance());
 
-        private EverythingIsLostException() {
-        }
+        final HazelcastAggregationRepository repoTwo
+                = new HazelcastAggregationRepository(REPO_NAME, false, getSecondInstance());
 
-        private EverythingIsLostException(String message) {
-            super(message);
-        }
+        repoOne.setUseRecovery(false);
+        repoTwo.setUseRecovery(false);
 
-        private EverythingIsLostException(String message, Throwable cause) {
-            super(message, cause);
-        }
+        final int completionSize = 4;
+        final String correlator = "CORRELATOR";
+        RouteBuilder rbOne = new RouteBuilder() {
+            @Override
+            public void configure() throws Exception {
 
-        private EverythingIsLostException(Throwable cause) {
-            super(cause);
-        }
+                from(DIRECT_ONE).routeId("AggregatingRouteOne")
+                        .aggregate(header(correlator))
+                        .aggregationRepository(repoOne)
+                        .aggregationStrategy(new SumOfIntsAggregationStrategy())
+                        .completionSize(completionSize)
+                        .to(MOCK_GOTCHA);
+            }
+        };
 
-        // not in jdk6
-        //        private EverythingIsLostException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        //            super(message, cause, enableSuppression, writableStackTrace);
-        //        }
+        RouteBuilder rbTwo = new RouteBuilder() {
+            @Override
+            public void configure() throws Exception {
+
+                from(DIRECT_TWO).routeId("AggregatingRouteTwo")
+                        .aggregate(header(correlator))
+                        .aggregationRepository(repoTwo)
+                        .aggregationStrategy(new SumOfIntsAggregationStrategy())
+                        .completionSize(completionSize)
+                        .to(MOCK_GOTCHA);
+            }
+        };
+
+        context().addRoutes(rbOne);
+        context().addRoutes(rbTwo);
+        context().start();
+
+        mock.expectedMessageCount(1);
+        mock.expectedBodiesReceived(1 + 2 + 3 + 4);
+
+        produceOne.sendBodyAndHeader(1, correlator, correlator);
+        produceTwo.sendBodyAndHeader(2, correlator, correlator);
+        produceOne.sendBodyAndHeader(3, correlator, correlator);
+        produceOne.sendBodyAndHeader(4, correlator, correlator);
+
+        mock.assertIsSatisfied();
     }
+
 }
