+++ test-infra/camel-test-infra-hazelcast/pom.xml
@@ -24,13 +24,14 @@
         <artifactId>camel-test-infra-parent</artifactId>
         <groupId>org.apache.camel</groupId>
         <relativePath>../camel-test-infra-parent/pom.xml</relativePath>
-        <version>3.12.0-SNAPSHOT</version>
+        <version>3.14.0-SNAPSHOT</version>
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
-
     <artifactId>camel-test-infra-hazelcast</artifactId>
+
     <name>Camel :: Test Infra :: Hazelcast</name>
+    <description>Hazelcast test infrastructure for Camel</description>
 
     <dependencies>
         <dependency>
